librbd: helper utility to retrieve config from the MON config store
A special "config://" URI prefix can be used to denote configuration
settings that should be (securely) pulled from the MON config store.
This will be first used in a follow-up commit to support storing
the S3 access and secret keys in the MON config store.

This "config://" syntax is already in-use by RGW and ceph-iscsi for
pulling secrets when deployed via cephadm.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>