osd: osdmap: do not assert target_v when encoding in stretch mode
Again, this was an overly-strict safety check. It passed testing against
(even old) userspace clients, but the kernel clients don't have all the
same feature bits and so when encoding for them, we see lower target_v values
and assert out erroneously.

As before, monitors and OSDs that actually need to interpret this data
are gated on cluster entry, so we don't need this safety check here.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>