crimson/os/seastore: validate parent immediately prior to adding pin
Mostly, operating on an invalid extent is harmless by construction --
it'll be caught by the read set check in submit_transaction.  However,
prior to adding an extent to the lba pin set, we need to check that it's
really still live, which we can do by checking its immediate parent.
Replacing the target extent would necessarily have involved a mutation
to the parent, which would have invalidated it.

Signed-off-by: Samuel Just <sjust@redhat.com>