doc/dev: improving teuthology docs
This commit is a proof of concept and a repair of
links to the "Integration Tests" section of the
Developer Guide.

This fixes the "reading a standalone test" link
from the "Running Tests in the Cloud" page
to the "Reading a Standalone Test" section in
the "Tests: Integration Tests" page.

Signed-off-by: Zac Dover <zac.dover@gmail.com>