mgr/volumes: deploy MDSs when creating fs
This line was (inadvertantly?) removed in 607537819efbf0d70662361e7ac35e5f16cd53c7
with the addition of the mds_autoscaler module.  However, that module
does not know how to deal with a filesystem that doesn't have a service
spec defined yet.  (Also, the module is not enabled by default.)

Signed-off-by: Sage Weil <sage@newdream.net>