common/Timer: add SafeTimer::add_event_at(real_clock::time_point)
so we can schedule an event with real_clock::time_point. timer should
wait using a mono_time, but user might want to pass a
real_clock::time_point to it when, for instance, scheduling a task which
should be performed in future at a specified time. in this case, we need
to convert the real_clock::time_point to a "known" time_point.

Signed-off-by: Kefu Chai <kchai@redhat.com>