rpm: refrain from using %make_build macro on SUSE
The recently merged commit 7136961e3c9ab7d8f4f463b5e3fd49f675c3a66d replaced

    make %{_smp_mflags}

with

    %make_build

for the stated purpose of hiding the %_smp_mflags macro in a higher-level macro.

But, on SUSE, the higher-level macro (%make_build) expands to:

    make -O %{_smp_mflags}

The addition of the -O flag makes the build considerably slower and increases
the memory requirement. The exact reason for this is unknown - possibly it's due
to a bug in make, although the same slowness was observed with ninja as well.
In any event, this is a deal-breaker when building in the OBS, because the build
infrastructure there is optimized for builds that do not require huge amounts of
memory and we would rather have a fast build with mixed up compiler messages
than a very slow one with synced compiler messages.

Fixes: 7136961e3c9ab7d8f4f463b5e3fd49f675c3a66d
Signed-off-by: Nathan Cutler <ncutler@suse.com>