cephfs: provide additional volume details on Windows
At the moment, the Windows volume name and serial number are hardcoded.

This change makes the volume name configurable, defaulting to "Ceph"
or "Ceph - <fs_name>". This makes it easier to identify Ceph mounts.

At the same time, we're going to retrieve the Ceph filesystem identifier
instead of using the hardcoded value.

Fixes: https://tracker.ceph.com/issues/49662
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>