cephfs-mirror: adjust PeerReplayer::cleanup_remote_dir() for purging any remote (sub)directory
Right now, the only path to cleanup (purge) is the remote file
system directory root (for configured snapshot directories).
With incremental sync, we would need the ability to cleanup
selective directories under configured snapshot directories.
Furthermore, the path passed to the function ise used to checking
if the mirror daemon needs to backoff due to mirroring being
disabled, daemon shutdown etc.

So, adjust the function prototype to accept the directory root
and the relative path to be cleaned up under this directory.

Signed-off-by: Venky Shankar <vshankar@redhat.com>