test/crimson/seastore: disable concurrent gc
The concurrent gc configuration is simpler for the most part, but
greatly complicates these unit tests as they might hit an eagain or not
based on gc interleaving.  For the purposes of these unit tests (and
other basically functional tests), we'll instead disable the background
behavior and invoke it directly after each submitted transaction.

Signed-off-by: Samuel Just <sjust@redhat.com>