crimson/monc: discard active/pending connections when reopening
Otherwise pending_conns vector just keeps growing with redundant
connections all trying to reach the same set of monitors.  When one
of the attempts finally succeeds, _finish_auth() will pick the first
connection with a matching entity_addr_t, designate it as active and
close all others.  The match is very likely to be wrong and hence the
actual authenticated connection gets closed, leaving the OSD with
a bogus active_con and no monitor session.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>