osd/PeeringState: init last_require_osd_release in ctor
so we can use the latest osdmap to tell what is the required osd release in
*current* cluster before consuming a new osdmap which is fed to
PeeringState only if we advance to a new map.

a osd shares its pg_info using MOSDPGInfo2 or MOSDPGInfo according to
PeeringState::last_require_osd_release, this happens before a new osdmap
is fully consumed.

Signed-off-by: Kefu Chai <kchai@redhat.com>