common/pick_address: Allow binding on loopback iface
in ee11ad8146161cf368513598411954cabe383351, "lo" is also skipped when
daemon is trying to find an address to bind. but that change reverts the
fix of 51e3a428dfa998548189297fe646cb936e9c86be, to address the problem.

an option named "ms_bind_exclude_lo_iface" is added, it defaults to
"true". but it can be changed to false to allow daemon to bind on "lo".

Fixes: https://tracker.ceph.com/issues/50012
Signed-off-by: Kefu Chai <kchai@redhat.com>