osd: osdmap: do not assert target_v when encoding under stretch mode
As in b5e904051893d11a1fe2d9879bf9c1aec366ca9e for the OSDMap, this
caused crashes when encoding for kernel clients, and is unnecessary
for servers because they are separately gated.

I did a full audit of every instance of "assert" I added to the codebase
to make sure this is the very last one of these issues.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>