osd: allow osd_client_message_size throttle set to zero
Once we set osd_client_message_size_cap to non-zero, it cannot be set
back to zero unless we set conf file to 0 and restart the osd.
If we don't want to limit messages by throttle anymore, we have to set
a great number as upper bound. So allow it change to 0.

Signed-off-by: haoyixing <haoyixing@kuaishou.com>