msg/msg_types: pass TYPE_DEFAULT as default_type to entity_addr_t::parse()
TYPE_NONE equals to 0, and we use it for the default value of entity's
messenger protocol version, if it is not specified. but we always use
TYPE_DEFAULT for the protocol version if it's not specified in the
parsed string or the parameter.

so we can just simplify the logic by assigning it to `newtype`, and
assign the parse value to `new_type` for better readability.

Signed-off-by: Kefu Chai <kchai@redhat.com>