cephadm: enable log to journald by default
Currently, when deployed with cephadm and podman, logs from ceph daemons are
routed as stderr -> conmon -> journald. We can send logs to journald directly
by connecting to its socket. And this has the following advantages:

- record structured metadata along with log message
  - prettier output from journalctl
    - no duplicated timestamp
    - colorized according to priority
    - multi-line logs are indented properly
  - easier to filter the logs afterward
- theoretically better performance
- workaround https://tracker.ceph.com/issues/49551

This depends on commit ca3202c. And depends on commit b4376f8 because cgroup is
needed to associate these logs to the systemd unit.

Signed-off-by: 胡玮文 <huww98@outlook.com>