test/cls_cas: allow multi hobjects tracked by cls_cas
in f263bb4cb685aff878ece561316aa642cba3f7d2, we started to use multiset
for tracking the references of hobject for snapshot support. as the same
hobject maps to multiple snapshots. and we don't want to consider
different snapshots as the same entry tracked by cls_cas.

but cls_cas.dup_get() still tries to verify that the `get` operation
is able to dedup the same referenced "source". but this does not apply
to "by_object" trunk ref type anymore.

since we cannot check/choose the chunk ref type used by OSD from the
client of the cls_cas, in this change, cls_cas.dup_get() is updated
to adapt the change solely for "by_object". otherwise we could skip
this test for "by_object" type and/or define another test for other
chunk ref types.

Fixes: https://tracker.ceph.com/issues/50339
Signed-off-by: Kefu Chai <kchai@redhat.com>