mon/OSDMonitor: prepare_new_pool() starts out with 1 pg
Starts a new pool with 1 pg if the autoscale mode is `on`.
If user didn't explicitly specify autoscale mode to be `on`,
we look at the default_autoscale_mode value that is set globally,
if that is `on` then we also start new pool with 1 pg. Else,
just use the default pg value.

Updated `common/options/global.yaml.in` to document the change.

Fixes: https://tracker.ceph.com/issues/49364

Signed-off-by: Kamoltat <ksirivad@redhat.com>