cmake: Revert "cmake: use file(GLOB ..) to find .yaml.in files"
This reverts commit 319cb5d3c95517ddcc54747ebe9426a3d2d549bc.

the list of .yaml.in file might change over time before we finish the
.yaml.in file split, but cmake would fail to figure out the list without
rerunning "cmake", so when a new .yaml.in file is introduced, developer
might end up with a FTBFS after pulling the change from remote repo.

so, we need to revert the file(GLOB ..) change, until all .yaml.in file
are created.

Signed-off-by: Kefu Chai <kchai@redhat.com>