librbd/api: stop at the first matched image
if compiled with libc++ instead of libstdc++, there is chance that
std::experimental::erase_if() is missing, so use a homebrew version
instead. we will be able to drop this change once we have the
luxury to use C++20.

also, as we only care about the first moving image whose name matches with the
specified one. let's take this opportunity to just stop at the first
image with the specified name.

this change trade a for loop with `find_if()` for better readability

Signed-off-by: Kefu Chai <kchai@redhat.com>