qa/tasks/cephfs/test_sessionmap: reap connections immediately
We have to reap connections promptly for this test to work.

This test was broken indirectly by d7a572d5507e72eaa21736e0689453006fb82cb7,
which moved the counter decrement to reap time instead of mark_down/stop
time.

The reaping is asynchronous, so allow for a delay in the count change.

Fixes: https://tracker.ceph.com/issues/50622
Signed-off-by: Sage Weil <sage@newdream.net>