mon/OSDMonitor: drop stale failure_info even if can_mark_down()
in 144e40b3e6700313f6e5ec23fed7cb93bba57ba3, we add a check to drop
stale failure_info reports. but if osdmap does not prohibit us from
marking the osd in question down, the branch checking the stale info
is not executed. in general, it is allowed to mark an osd down, so
the fix of 144e40b3e6700313f6e5ec23fed7cb93bba57ba3 just fails to
work.

in this change, we check for stale failure report of osd in question
as long as the osd is not marked down in the same function. this should
address the slow ops of failure report issue.

Fixes: https://tracker.ceph.com/issues/50964
Signed-off-by: Kefu Chai <kchai@redhat.com>