crimson/osd: wait for SIGINT and SIGTERM before stopping
this change addresses an regression introduced by
e38b84301b354c6d62e943d2f4f7c77c1b814885. as we should not stop
services without being asked to do so.

in this change, signal handler for SIGINT and SIGTERM is registered to
handle these signals, and in the seastar thread, we wait until any of
these two signals is caught.

Signed-off-by: Kefu Chai <kchai@redhat.com>