qa/standalone: Use osd op queue = wpq in activate_osd()
This change is a follow-up to commit
2a6eb367fe31c373b84f3fe2c33bfdfc4ca7bdf6 that set the scheduler to wpq in
run_osd() and run_osd_filestore(). In addition, activate_osd() too has to
set the scheduler type to 'wpq' in order to be consistent and avoid test
failures.

The above is a temporary measure until all the standalone tests are
modified to run well with the mclock_scheduler.

Fixes: https://tracker.ceph.com/issues/51074
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>