include/cpp-btree: drop node handle support
this is a leftover in ff4dbabefec0237494990c1ae7d0421f0db55697,

the node support is useful if we want to move a node around in a
node-based container. but in the case of b-tree, multiple elements
are packed into a single node/block, so we cannot benefit from this
feature, i think. so just drop all of them.

Signed-off-by: Kefu Chai <kchai@redhat.com>