cmake: require CMake v3.16
since we dropped the support of bionic, we now have the luxury of using
newer CMake! and by using CMake 3.16, we now have access to

* add_compile_definitions(), which was introduced in cmake 3.12

also, drop the comment of

> remove cmake/modules/FindPython* once 3.12 is required

because we need to support python 3.10 in future, better off keeping
the FindPython files around, so we can backport cmake changes from
upstream when necessary.

we are allowed to bump up the CMake's required to 3.16, because:

- ubuntu focal ships with cmake 3.16.3
- RHEL/CentOS8's AppStream has cmake 3.18

Signed-off-by: Kefu Chai <kchai@redhat.com>