mon/MgrMonitor: drop nautilus modules from always_on_modules
the list of always_on_modules for the older versions is only for the
upgrade use case, where the monitor is still the old release, so
we have to keep an entry for the old release in the always_on_modules
map.

but since we don't upgrade from a release older than 2 releases, there
is no need to keep an entry for nautilus.

Signed-off-by: Kefu Chai <kchai@redhat.com>