crimson/os/seastore/extent_placement_manager: remove templated IndexT, HintT
extent_placement_manager is specifically where we want to be making choices
based on HintT, exposing internal details like IndexT or the HintT->IndexT
mapping is confusing.  Moreover, external users will need to be coded with
specific HintT values in mind, so let's just make it concrete in
seastore_types.h.  We can make this more complicated when there's actual
logic depending on it.