cmake: make rgw_common a static library
it was changed from a shared library to object library back in
1d3837e578eac9e7ff44781d623b16feaa258a41, because we wanted to link both
rgw_a and rgw_admin_user against it.

but now that rgw_admin_user was dropped in
dc1947ff396277dea923b365e2342a23ca865414, let's compile rgw_common
as a static library. it'd be much easier for us to apply various compile
options and linkages to it this way.

Signed-off-by: Kefu Chai <kchai@redhat.com>