debian/control: make cryptsetup-bin a ceph-volume runtime dep
was originally added by 61993ea8a52f1be091657b2754b3a6919b142fda and
dfd19454386cb57bf6b87f701d603b4d51cdefb6 as a dependency of ceph-disk.
although, ceph-disk was replaced by ceph-volume, cryptsetup-bin is still
used.

but it should be a runtime dependency of ceph-volume. and it is not a
build dependency of ceph, so it's removed from the build dependency list
as well.

Signed-off-by: Kefu Chai <kchai@redhat.com>